add ClassifiedTarget type to core
